Use ARAL for labels (#20502)
* Use ARAL for labels

* Change label version to uint32_t

* During shutdown call rrdlabels_aral_destroy() when running with FSANITIZE_ADDRESS

* Remove memory accounting (handled by ARAL)

* Fix unittest

* Fix unit tests part 2

(cherry picked from commit 4f3397ef94ffaba0131a56fd1348f5714e15676c)

+++ src/database/sqlite/sqlite_metadata.c
@@ -840,8 +840,8 @@ static int chart_label_store_to_sql_callback(const char *name, const char *value
 
 static int check_and_update_chart_labels(RRDSET *st, BUFFER *work_buffer)
 {
-    size_t old_version = st->rrdlabels_last_saved_version;
-    size_t new_version = rrdlabels_version(st->rrdlabels);
+    uint32_t old_version = st->rrdlabels_last_saved_version;
+    uint32_t new_version = rrdlabels_version(st->rrdlabels);
 
     if (new_version == old_version)
         return 0;
